What: work for go1.7 and !go1.7
Why:

  * use golang.org/x/net/context in !go1.7

How:

  * nop
